Remove listeners only when there is an overflow
There was a bug and we removed listeners even when the buffer was not
full. That meant we ended tracing earlier than necessary in
non-streaming case.

Also fixed a bug where we were passing uint16_t for storing a 24-bit
value which was uncovered when adding a test for this (ag/36025507)

Bug: 441450730
Test: art/test.py
Flag: EXEMPT BUGFIX

+++ runtime/trace.h
@@ -152,7 +152,7 @@ static inline void Append2LE(uint8_t* buf, uint16_t val) {
 }
 
 // TODO: put this somewhere with the big-endian equivalent used by JDWP.
-static inline void Append3LE(uint8_t* buf, uint16_t val) {
+static inline void Append3LE(uint8_t* buf, uint32_t val) {
   *buf++ = static_cast<uint8_t>(val);
   *buf++ = static_cast<uint8_t>(val >> 8);
   *buf++ = static_cast<uint8_t>(val >> 16);
